finishing practice5
the last issue with this function is that i want to return the vector, but since i've made a new vector on the heap, i do not know how to return the entire vector that is on the heap.

so far i am returning the first elemnet within the vector. need to figure out how to return a  new vector that is made on the heap back to main and then need to figure out how to delete said new vector that was made on the heap.

then i will be completely done with practice 5 vector and move onto the example prompts that were given on csuf website